+++ src/service/BootstrapUtils.ts
@@ -32,7 +32,7 @@ import * as Handlebars from 'handlebars';
 import { get } from 'https';
 import * as _ from 'lodash';
 import { platform, totalmem } from 'os';
-import { basename, join } from 'path';
+import { basename, join, resolve } from 'path';
 import { Convert, Deadline, DtoMapping, LinkAction, NetworkType, Transaction, UInt64, VotingKeyLinkTransaction } from 'symbol-sdk';
 import * as util from 'util';
 import { LogType } from '../logger';
@@ -255,6 +255,14 @@ export class BootstrapUtils {
         }
     }
 
+    public static resolveRootFolder(): string {
+        const rootFolder = resolve(__dirname, '../..');
+        if (!existsSync(join(rootFolder, 'presets', 'shared.yml'))) {
+            throw new Error(`Root Folder ${rootFolder} does not look right!`);
+        }
+        return rootFolder;
+    }
+
     public static async runImageUsingExec({
         catapultAppFolder,
         image,
@@ -373,6 +381,7 @@ export class BootstrapUtils {
                         } else {
                             await fsPromises.copyFile(fromPath, destinationFile);
                         }
+                        await fsPromises.chmod(destinationFile, 0o600);
                     }
                 } else if (stat.isDirectory()) {
                     await fsPromises.mkdir(toPath, { recursive: true });